type_traits: Move to public module and add documentation
