+++ src/stdgpu/atomic.cuh
@@ -31,9 +31,9 @@
 #include <type_traits>
 
 #include <stdgpu/execution.h>
-#include <stdgpu/impl/type_traits.h>
 #include <stdgpu/memory.h>
 #include <stdgpu/platform.h>
+#include <stdgpu/type_traits.h>
 
 namespace stdgpu
 {
