+++ src/stdgpu/cuda/atomic.cuh
@@ -18,7 +18,7 @@
 
 #include <type_traits>
 
-#include <stdgpu/impl/type_traits.h>
+#include <stdgpu/type_traits.h>
 
 namespace stdgpu::cuda
 {
