+++ src/stdgpu/execution.h
@@ -23,7 +23,7 @@
 
 #include <type_traits>
 
-#include <stdgpu/impl/type_traits.h>
+#include <stdgpu/type_traits.h>
 
 /**
  * \file stdgpu/execution.h
